+++ CHANGELOG.md
@@ -14,6 +14,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added implementation for cucumber tests for quality investigations
 - Separation of auto complete mechanism (selected / searched elements)
 - Added new step definition for cucumber tests "I use assets with ids {string}" allowing to specify assets used for notification creation
+- Emit change check to observables in frontend
 
 ### Changed
 - Filter configuration for tables to be resuable and easy to adapt
@@ -27,6 +28,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Cucumber test steps for creating notifications no longer support default assetId when no asset is provided with previous step
 - Upgraded the Upload_Testdata job in Argo Workflow to fix bugs
 - Auto format for frontend source code applied
+- Default pagination size to 50.
 
 ### Removed
 - removed asset filters ( qualityInvestigationIdsInStatusActive, qualityInvestigationIdsInStatusActive )